// BookMyMovie/src/app/Models/paymentUrl.ts
export class paymentUrl{
    showId: String;
    movieId: String;
    theatreId: String;
    seats: String[];
    totalseat: Number;
    showtime: String;
    date: String;

//  constructor(showId: String,
//     movieId: String, theatreId: String, seats: String[], 
//     totalseat: Number, showtime: String, date: String)
//         {
        
//             this.showId = showId;
//             this.movieId = movieId;
//             this.theatreId = theatreId;
//             this.seats = seats;
//             this.totalseat = totalseat;
//             this.showtime = showtime;
//             this.date = date;
//     }
}